+++ qol-bundle/src/init.lua
@@ -39,6 +39,13 @@ QOL_BUNDLE.init = function(self)
   -- we can use the value directly as the index
   self.wheel_of_fortune_current_option = self.config.wheel_of_fortune_value
   self.wheel_of_fortune_current_option_val = self.config.wheel_of_fortune_value
+
+  -- eight_ball_joker_value is 4 by default
+  self.config.eight_ball_joker_value = self.config.eight_ball_joker_value or 4
+  -- Since our values match their indices in the options array {1, 2, 3, 4},
+  -- we can use the value directly as the index
+  self.eight_ball_joker_current_option = self.config.eight_ball_joker_value
+  self.eight_ball_joker_current_option_val = self.config.eight_ball_joker_value
 end
 
 -- Run the initialization function